"correct" but "incorrect"
there is a problem with the recurse function so it will need more work.
got the bottom up method working but it is getting different results
from the expected output. how the method currently works it is getting
different denomination values from the expected but it has the same
total amount of denominations used?