more work on assignment 4
something weird is going on with the algorithm for the edit distance.
when it starts to get negative values the values are not what they are
supposed to be in the expected output but the positive costs seem to be
fine. the algorithm in the presentation is different than what I was
expecting and it to a minute to figure out what the problem was. I
changed it to a similar structure as the myers code I wrote for the
project and that seemed to help.