finished tree_vc
finished the tree vertex cover algorithm. I used the code that I had
previously written for the assignment but had to make changes because
the graph was expected to be undirected vs directed (probably missed
that durring lecture). also made an iterative and recursive version
because I thought the recursive version was stackoverflowing for the
wrong reason. it was entering an infinte loop because I was not tracking
if I had visited the node already.